# packages/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DECODED API",
    "description": "AI 기반 미디어 디스커버리 플랫폼 API",
    "contact": {
      "name": "DECODED Team"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://dev.decoded.style",
      "description": "Development server"
    },
    {
      "url": "https://api.decoded.style",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "통합 검색",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "검색어",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "카테고리 필터 (fashion, living, tech, beauty)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "media_type",
            "in": "query",
            "description": "미디어 타입 필터 (drama, movie, mv, youtube, variety)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "context",
            "in": "query",
            "description": "컨텍스트 필터 (airport, stage, mv, red_carpet 등)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "has_adopted",
            "in": "query",
            "description": "채택된 Solution이 있는 Spot만 필터링",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "정렬 방식 (recent, popular, solution_count, default: relevant)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호 (default: 1)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "페이지당 개수 (default: 20, max: 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "검색 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Facets": {
        "type": "object",
        "description": "Facets (필터링 집계)",
        "properties": {
          "category": {
            "type": [
              "object",
              "null"
            ],
            "description": "카테고리별 개수",
            "additionalProperties": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "context": {
            "type": [
              "object",
              "null"
            ],
            "description": "컨텍스트별 개수",
            "additionalProperties": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "media_type": {
            "type": [
              "object",
              "null"
            ],
            "description": "미디어 타입별 개수",
            "additionalProperties": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
      "MediaSource": {
        "type": "object",
        "description": "미디어 소스 정보",
        "required": [
          "type"
        ],
        "properties": {
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "미디어 제목"
          },
          "type": {
            "type": "string",
            "description": "미디어 타입"
          }
        }
      },
...
      "PaginationMeta": {
        "type": "object",
        "required": [
          "current_page",
          "per_page",
          "total_items",
          "total_pages"
        ],
        "properties": {
          "current_page": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_items": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "description": "검색 응답",
        "required": [
          "data",
          "facets",
          "pagination",
          "query",
          "took_ms"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "description": "검색 결과 목록"
          },
          "facets": {
            "$ref": "#/components/schemas/Facets",
            "description": "Facets (필터링 집계)"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta",
            "description": "페이지네이션 정보"
          },
          "query": {
            "type": "string",
            "description": "검색어"
          },
          "took_ms": {
            "type": "integer",
            "format": "int64",
            "description": "검색 소요 시간 (ms)",
            "minimum": 0
          }
        }
      },
      "SearchResultItem": {
        "type": "object",
        "description": "검색 결과 항목",
        "required": [
          "id",
          "type",
          "image_url",
          "spot_count",
          "view_count"
        ],
        "properties": {
          "artist_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "아티스트명"
          },
          "context": {
            "type": [
              "string",
              "null"
            ],
            "description": "컨텍스트"
          },
          "group_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "그룹명"
          },
          "highlight": {
            "type": [
              "object",
              "null"
            ],
            "description": "하이라이트 (검색어 강조)",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Post ID"
          },
          "image_url": {
            "type": "string",
            "description": "이미지 URL"
          },
          "media_source": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MediaSource",
                "description": "미디어 소스"
              }
            ]
          },
          "spot_count": {
            "type": "integer",
            "format": "int32",
            "description": "Spot 개수"
          },
          "type": {
            "type": "string",
            "description": "결과 타입 (post)"
          },
          "view_count": {
            "type": "integer",
            "format": "int32",
            "description": "조회수"
          }
        }
      },
...
    }
  },
...
}